# swagger/proto/user.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "proto/user.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "Users"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/v1/register": {
      "post": {
        "operationId": "Users_Register",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/userResponseUserData"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/userUserRegister"
            }
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users": {
      "get": {
        "operationId": "Users_List",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/userResponseUserData"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "Users"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "userResponseUserData": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/userUserWithoutPassword"
          }
        }
      }
    },
    "userUserRegister": {
      "type": "object",
      "properties": {
        "first_name": {
          "type": "string",
          "title": "int64 id = 1;",
          "required": [
            "first_name"
          ]
        },
        "last_name": {
          "type": "string",
          "required": [
            "last_name"
          ]
        },
        "email": {
          "type": "string",
          "required": [
            "email"
          ]
        },
        "password": {
          "type": "string",
          "required": [
            "password"
          ]
        },
        "phone": {
          "type": "string",
          "required": [
            "phone"
          ]
        }
      },
      "required": [
        "first_name",
        "last_name",
        "email",
        "password",
        "phone"
      ]
    },
    "userUserRole": {
      "type": "string",
      "enum": [
        "Student",
        "Admin"
      ],
      "default": "Student"
    },
    "userUserWithoutPassword": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "first_name": {
          "type": "string"
        },
        "last_name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "role": {
          "$ref": "#/definitions/userUserRole"
        }
      }
    }
  }
}